requirements.txt
dateparser==0.7.0
requests
websocket-client==0.54.0
bitmex
python-dotenv
raven==6.9.0
